+++ src/lib/components/fund/jsonFund.tsx
@@ -1,0 +1,10 @@
+import JsonInput from "lib/components/json/JsonInput";
+
+interface JsonFundProps {
+  setValue: (value: string) => void;
+
+  assetsJson: string;
+}
+export const JsonFund = ({ setValue, assetsJson }: JsonFundProps) => (
+  <JsonInput text={assetsJson} setText={setValue} height="160px" />
+);